Added config file options for continuing in the dialplan
A system admin is now able to specify a context, extension,
and priority so that when a user presses the DTMF digit 3
it kicks the user out of the bridge and runs through the
specified dialplan extension. If a dialplan context or extension
isn't specified in the config.json then the user will be thrown
out of the bridge and whatever comes after the stasis app in
the dialplan will be run.